requirements.txt
pymongo
requests
requesocks
gevent
jieba
wordcloud
Pillow